Add DSAbstractAction to make anonymous actions easier to implement.
